Moved the function that converts a spatial model to a model.
