+++ stochss/handlers/util/stochss_model.py
@@ -341,22 +341,6 @@ class StochSSModel(StochSSBase):
         return g_model
 
 
-    def convert_to_model(self):
-        '''
-        Convert a spatial model to a non_spatial model
-
-        Attributes
-        ----------
-        '''
-        if self.model is None:
-            s_model = self.load()
-        s_model['is_spatial'] = False
-        m_path = self.path.replace(".smdl", ".mdl")
-        m_file = self.get_file(path=m_path)
-        message = f"{self.get_file()} was successfully convert to {m_file}!"
-        return {"Message":message, "File":m_file}, {"model":s_model, "path":m_path}
-
-
     def convert_to_sbml(self):
         '''
         Convert a model to a SBML model
